+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Don't allow `BackendModification`s to change resource type
 - Add event allocation functions directly taking `BackendModification`s.
+- Move from `MonadCleanup` to `general-allocate`
 
 ## 0.5.0.0 -- 2022-10-23